// frontend/src/app/modules/candidate/interfaces/IExperience.ts
export interface IExperience {
    experienceGuid: string;
    employmentTypeGuid: string;
    employmentTypeName: string;
    locationTypeGuid: string;
    locationTypeName: string;
    designationGuid: string;
    designation: string;
    companyGuid: string;
    companyName: string;
    country: string;
    state: string;
    city: string;
    experienceStartDate: string;
    experienceEndDate: string;
    isExperienceActive: boolean;
    companyDescription: string;
    experienceDescription: string;
}